infra/requirements.txt

beautifulsoup4
lxml
urllib3
sendgrid
python-dotenv
openai
unidecode
google-api-python-client
sqlalchemy